working...

+++ web/db/myDb.sql
@@ -4,31 +4,31 @@
 
 -- Create tables and sequences --
 CREATE TABLE SYSTEM_USER
-( system_user_id   INT
+( system_user_id   BIGINT
 , system_user_name TEXT CONSTRAINT nn_system_user_1 NOT NULL
 , CONSTRAINT pk_system_user_1 PRIMARY KEY (system_user_id)
 , CONSTRAINT uk_system_user_1 UNIQUE (system_user_name));
 CREATE SEQUENCE system_user_s1 START WITH 1001;
 
 CREATE TABLE COMMON_LOOKUP
-( common_lookup_id    INT
+( common_lookup_id    BIGINT
 , common_lookup_table TEXT CONSTRAINT nn_common_lookup_1 NOT NULL
 , common_lookup_type  TEXT CONSTRAINT nn_common_lookup_2 NOT NULL
-, created_by          INT
+, created_by          BIGINT
 , CONSTRAINT pk_common_lookup_1 PRIMARY KEY (common_lookup_id)
 , CONSTRAINT fk_common_lookup_1 FOREIGN KEY (created_by) REFERENCES system_user(system_user_id)
 , CONSTRAINT uk_common_lookup_1 UNIQUE (common_lookup_table, common_lookup_type));
 CREATE SEQUENCE common_lookup_s1 START WITH 1001;
 
 CREATE TABLE PERSON
-( person_id    INT
-, person_type  INT
+( person_id    BIGINT
+, person_type  BIGINT
 , name_first   TEXT
 , name_middle  TEXT
 , name_last    TEXT
 , brithday     DATE
 , picture_url  TEXT
-, created_by   INT CONSTRAINT nn_person_1 NOT NULL
+, created_by   BIGINT CONSTRAINT nn_person_1 NOT NULL
 , CONSTRAINT pk_person_1 PRIMARY KEY (person_id)
 , CONSTRAINT fk_person_1 FOREIGN KEY (created_by)  REFERENCES system_user (system_user_id)
 , CONSTRAINT fk_person_2 FOREIGN KEY (person_type) REFERENCES common_lookup (common_lookup_id)
@@ -36,13 +36,13 @@ CREATE TABLE PERSON
 CREATE SEQUENCE person_s1 START WITH 1001;
 
 CREATE TABLE TELEPHONE
-( telephone_id     INT
-, telephone_type   INT
-, person_id        INT
-, country_code     INT
-, area_code        INT CONSTRAINT nn_telephone_1 NOT NULL
-, telephone_number INT CONSTRAINT nn_telephone_2 NOT NULL
-, created_by       INT
+( telephone_id     BIGINT
+, telephone_type   BIGINT
+, person_id        BIGINT
+, country_code     BIGINT
+, area_code        BIGINT CONSTRAINT nn_telephone_1 NOT NULL
+, telephone_number BIGINT CONSTRAINT nn_telephone_2 NOT NULL
+, created_by       BIGINT
 , CONSTRAINT pk_telephone_1 PRIMARY KEY (telephone_id)
 , CONSTRAINT fk_telephone_1 FOREIGN KEY (person_id)      REFERENCES person (person_id)
 , CONSTRAINT fk_telephone_2 FOREIGN KEY (created_by)     REFERENCES system_user (system_user_id)
@@ -50,15 +50,15 @@ CREATE TABLE TELEPHONE
 CREATE SEQUENCE telephone_s1 START WITH 1001;
 
 CREATE TABLE ADDRESS
-( address_id     INT
-, address_type   INT
-, person_id      INT
+( address_id     BIGINT
+, address_type   BIGINT
+, person_id      BIGINT
 , country        TEXT
 , state          TEXT
 , postal_code    TEXT
 , street_address TEXT
 , apartment      TEXT
-, created_by     INT
+, created_by     BIGINT
 , CONSTRAINT pk_address_1 PRIMARY KEY (address_id)
 , CONSTRAINT fk_address_1 FOREIGN KEY (person_id)    REFERENCES person (person_id)
 , CONSTRAINT fk_address_2 FOREIGN KEY (created_by)   REFERENCES system_user (system_user_id)
@@ -66,11 +66,11 @@ CREATE TABLE ADDRESS
 CREATE SEQUENCE address_s1 START WITH 1001;
 
 CREATE TABLE EMAIL
-( email_id   INT
-, email_type INT
-, person_id  INT
+( email_id   BIGINT
+, email_type BIGINT
+, person_id  BIGINT
 , email      TEXT
-, created_by INT
+, created_by BIGINT
 , CONSTRAINT pk_email_1 PRIMARY KEY (email_id)
 , CONSTRAINT fk_email_1 FOREIGN KEY (person_id)  REFERENCES person (person_id)
 , CONSTRAINT fk_email_2 FOREIGN KEY (created_by) REFERENCES system_user (system_user_id)
